RR-64 reorganize spline tests
