+++ TeamCode/src/main/java/com/acmerobotics/relicrecovery/path/parametric/SplinePath.java
@@ -46,14 +46,8 @@ public class SplinePath implements ParametricPath {
         }
 
         computeLength();
-        System.out.println("input start: " + startPose + ", end: " + endPose);
 
         headingFlipped = Math.abs(Angle.norm(startPose.heading() - start().heading())) > Math.PI / 2;
-
-        System.out.println("calc  start: " + start() + ", end: " + end());
-
-        System.out.println("heading offset: " + Math.toDegrees(headingOffset));
-        System.out.println();
     }
 
     private double valueAt(double percentage) {
@@ -112,10 +106,11 @@ public class SplinePath implements ParametricPath {
 
         double x = knotDistance * percentage;
         double y = valueAt(percentage);
-        double heading = Angle.norm(Math.atan(derivative) + headingOffset);
+        double heading = Math.atan(derivative) + headingOffset;
         if (headingFlipped) {
             heading += Math.PI;
         }
+        heading = Angle.norm(heading);
 
         return new Pose2d(
                 x * Math.cos(headingOffset) - y * Math.sin(headingOffset) + xOffset,
