+++ TeamCode/src/test/java/com/acmerobotics/relicrecovery/path/SplinePathTest.java
@@ -8,9 +8,8 @@ import com.acmerobotics.relicrecovery.path.parametric.ParametricPath;
 import org.junit.Ignore;
 import org.junit.Test;
 
-import java.io.File;
-import java.io.FileWriter;
-import java.io.IOException;
+import static com.acmerobotics.relicrecovery.path.TrajectoryUtil.writeParametricPath;
+import static com.acmerobotics.relicrecovery.path.TrajectoryUtil.writePathMotionSegment;
 
 public class SplinePathTest {
     @Ignore
@@ -23,51 +22,5 @@ public class SplinePathTest {
 
         TrajectorySegment motionSegment = new ParametricSegment(spline);
         writePathMotionSegment(motionSegment, "simpleMotionSegment");
-    }
-
-    public static void writeParametricPath(ParametricPath path, String name) {
-        try {
-            File outputDir = new File("./out");
-            outputDir.mkdirs();
-            FileWriter writer = new FileWriter(new File(outputDir,name + ".csv"));
-            writer.write("t, x, y, theta, vx, vy, omega, ax, ay, alpha\n");
-            System.out.printf("writing path: length=%.2fin", path.length());
-            for (int i = 0; i <= 200; i++) {
-                double displacement = i / 200.0 * path.length();
-                Pose2d pose = path.getPose(displacement);
-                Pose2d velocity = path.getDerivative(displacement);
-                Pose2d acceleration = path.getSecondDerivative(displacement);
-                writer.write(String.format("%f, %f, %f, %f, %f, %f, %f, %f, %f, %f\n", displacement,
-                        pose.x(), pose.y(), pose.heading(),
-                        velocity.x(), velocity.y(), velocity.heading(),
-                        acceleration.x(), acceleration.y(), acceleration.heading()));
-            }
-            writer.close();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-
-    public static void writePathMotionSegment(TrajectorySegment motionSegment, String name) {
-        try {
-            File outputDir = new File("./out");
-            outputDir.mkdirs();
-            FileWriter writer = new FileWriter(new File(outputDir,name + ".csv"));
-            writer.write("t, x, y, theta, vx, vy, omega, ax, ay, alpha\n");
-            for (int i = 0; i <= 200; i++) {
-                double time = i / 200.0 * motionSegment.duration();
-                Pose2d pose = motionSegment.getPose(time);
-                Pose2d velocity = motionSegment.getVelocity(time);
-                Pose2d acceleration = motionSegment.getAcceleration(time);
-                writer.write(String.format("%f, %f, %f, %f, %f, %f, %f, %f, %f, %f\n", time,
-                        pose.x(), pose.y(), pose.heading(),
-                        velocity.x(), velocity.y(), velocity.heading(),
-                        acceleration.x(), acceleration.y(), acceleration.heading()));
-
-            }
-            writer.close();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
     }
 }
